chore : rename rest-api  , directory structure
