+++ src/main/java/com/Techeer/Team_C/domain/user/auth/JwtTokenProvider.java
@@ -1,6 +1,7 @@
 
 package com.Techeer.Team_C.domain.user.auth;
 
+import com.Techeer.Team_C.domain.user.dto.TokenDto;
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
