+++ src/main/java/com/Techeer/Team_C/domain/user/dto/UserDto.java
@@ -2,20 +2,12 @@ package com.Techeer.Team_C.domain.user.dto;
 //DB의 테이블구조와 직접적으로 연관된 entity를 사용하는것은 좋지 않기때문에 중간의 데이터 교환을 위한 Dto가 필요하다.
 // entity의 속성과 동일하게 작성한다.
 
-
 import lombok.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
+
 
 @Getter
 @Setter
