+++ src/main/java/com/Techeer/Team_C/domain/user/repository/RefreshTokenRepository.java
@@ -1,12 +1,16 @@
-package com.Techeer.Team_C.domain.user.auth;
+package com.Techeer.Team_C.domain.user.repository;
 
-import com.Techeer.Team_C.domain.user.auth.RefreshToken;
+import com.Techeer.Team_C.domain.user.entity.RefreshToken;
 import org.springframework.stereotype.Repository;
-//import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+
+/*
+해당 repositorys는 mysql이 아닌, 메모리기반의 repository입니다.
+향후 mysql과 연결해 직접 데이터를 가져오고 저장하는 기능을 구현할 예정입니다.
+ */
 
 @Repository
 public class RefreshTokenRepository{ //userId값으로 토큰을 가져올 때 사용
